+++ src/pages/common/report/ReportPaymentPage.tsx
@@ -1,11 +1,7 @@
-import { useEffect, useState } from 'react';
+import { useEffect } from 'react';
 import { FaArrowLeft } from 'react-icons/fa6';
-import { useNavigate, useParams } from 'react-router-dom';
+import { useNavigate } from 'react-router-dom';
 
-import {
-  convertReportPriceType,
-  useGetReportPriceDetail,
-} from '../../../api/report';
 import Card from '../../../components/common/report/Card';
 import Heading1 from '../../../components/common/report/Heading1';
 import Heading2 from '../../../components/common/report/Heading2';
@@ -13,7 +9,7 @@ import Label from '../../../components/common/report/Label';
 import BottomSheet from '../../../components/common/ui/BottomSheeet';
 import Input from '../../../components/common/ui/input/Input';
 import useReportProgramInfo from '../../../hooks/useReportProgramInfo';
-import useReportApplicationStore from '../../../store/useReportApplicationStore';
+import { ReportPaymentSection } from './ReportApplyPage';
 
 const ReportPaymentPage = () => {
   const navigate = useNavigate();
@@ -37,7 +33,7 @@ const ReportPaymentPage = () => {
       <main className="mb-8 flex flex-col gap-10">
         <ProgramInfoSection />
         <UsereInfoSection />
-        <PaymentSection />
+        <ReportPaymentSection />
       </main>
       <BottomSheet>
         <button
@@ -116,111 +112,6 @@ const UsereInfoSection = () => {
             가입한 이메일과 동일
           </label>
           <Input name="contactEmail" placeholder="example@example.com" />
-        </div>
-      </div>
-    </section>
-  );
-};
-
-const PaymentSection = () => {
-  const { reportId } = useParams();
-
-  const [priceInfo, setPriceInfo] = useState({
-    report: 0,
-    feedback: 0,
-    discount: 0,
-    coupon: 0,
-    total: 0,
-  });
-
-  const { data: reportApplication } = useReportApplicationStore();
-  const { data: reportPriceDetail } = useGetReportPriceDetail(Number(reportId));
-
-  // 결제 가격 계산
-  useEffect(() => {
-    const reportPriceInfo = reportPriceDetail?.reportPriceInfos?.find(
-      (info) => info.reportPriceType === reportApplication.reportPriceType,
-    );
-    const feedbackPriceInfo = reportPriceDetail?.feedbackPriceInfo;
-
-    const report = reportPriceInfo?.price as number;
-    const feedback = feedbackPriceInfo?.feedbackPrice as number;
-    let discount = 0;
-    let total = 0;
-
-    discount += reportPriceInfo?.discountPrice as number;
-    reportApplication.optionIds.forEach((optionId) => {
-      discount += reportPriceDetail?.reportOptionInfos?.find(
-        (info) => info.reportOptionId === optionId,
-      )?.discountPrice as number;
-    });
-
-    if (reportApplication.isFeedbackApplied)
-      discount += feedbackPriceInfo?.feedbackDiscountPrice as number;
-
-    total = report + feedback - discount;
-
-    setPriceInfo({
-      report,
-      feedback,
-      discount,
-      coupon: 0,
-      total,
-    });
-  }, [reportPriceDetail]);
-
-  return (
-    <section>
-      <Heading2>결제 정보</Heading2>
-      {/* <div className="mt-6 flex gap-2.5">
-        <Input
-          className="w-full"
-          type="text"
-          placeholder="쿠폰 번호를 입력해주세요."
-        />
-        <button className="shrink-0 rounded-sm bg-primary px-4 py-1.5 text-xsmall14 font-medium text-neutral-100">
-          쿠폰 등록
-        </button>
-      </div> */}
-      <hr className="my-5" />
-      <div className="flex flex-col">
-        <div className="flex h-10 items-center justify-between px-3 text-neutral-0">
-          <span>
-            서류 진단서 (
-            {reportApplication.optionIds.length === 0
-              ? convertReportPriceType(reportApplication.reportPriceType)
-              : `${convertReportPriceType(reportApplication.reportPriceType)} + 옵션`}
-            )
-          </span>
-          {/* 서류 진단 + 사용자가 선택한 모든 옵션 가격을 더한 값 */}
-          <span>{priceInfo.report.toLocaleString()}원</span>
-        </div>
-        <div className="flex h-10 items-center justify-between px-3 text-neutral-0">
-          <span>1:1 피드백</span>
-          {/* 1:1 피드백 가격 */}
-          <span>{priceInfo.feedback.toLocaleString()}원</span>
-        </div>
-        <div className="flex h-10 items-center justify-between px-3 text-neutral-0">
-          {/* 서류진단 + 사용자가 선택한 모든 옵션 + 1:1 피드백의 할인 가격을 모두 더한 값 */}
-          <span>
-            {Math.ceil(
-              (priceInfo.discount / (priceInfo.report + priceInfo.feedback)) *
-                100,
-            )}
-            % 할인
-          </span>
-          <span>-{priceInfo.discount.toLocaleString()}원</span>
-        </div>
-        {/* <div className="flex h-10 items-center justify-between px-3 text-primary">
-          <span>쿠폰할인</span>
-          <span className="font-bold">
-            -{priceInfo.coupon.toLocaleString()}원
-          </span>
-        </div> */}
-        <hr className="my-5" />
-        <div className="flex h-10 items-center justify-between px-3 font-semibold text-neutral-0">
-          <span>결제금액</span>
-          <span>{priceInfo.total.toLocaleString()}원</span>
         </div>
       </div>
     </section>
